+++ srl/base/run/play_mp_memory.py
@@ -45,6 +45,10 @@ class MpConfig:
     train_to_mem_queue_capacity: int = 100
     mem_to_train_queue_capacity: int = 5
 
+    # memory
+    return_memory_data: bool = False
+    return_memory_timeout: int = 60 * 60 * 1
+
 
 # --------------------
 # actor
@@ -167,6 +171,7 @@ def _run_actor(
     remote_board: Any,
     actor_id: int,
     end_signal: Any,
+    last_worker_param_queue: mp.Queue,
 ):
     # import faulthandler
     # faulthandler.enable()
@@ -227,6 +232,12 @@ def _run_actor(
         state.worker = workers[main_worker_idx]
         state.workers = workers
         core_play.play(context, state, callbacks=callbacks)
+
+        if context.rl_config.use_update_parameter_from_worker():
+            # actor0のみ送信
+            if actor_id == 0:
+                logger.info(f"[actor{actor_id}] send parameter data")
+                last_worker_param_queue.put(parameter.backup())
 
     except MemoryError:
         import gc
@@ -320,7 +331,9 @@ def _run_memory(
             [c.on_memory(cfg.context, info) for c in _calls_on_memory]
 
         [c.on_memory_end(cfg.context, info) for c in cfg.callbacks]
-        last_mem_queue.put(memory.backup(compress=True))
+        if cfg.return_memory_data:
+            logger.info("[memory] send memory data")
+            last_mem_queue.put(memory.backup(compress=True))
 
     except MemoryError:
         import gc
@@ -572,7 +585,6 @@ def _run_trainer(
             logger.info("[trainer] end")
         if not exception_queue.empty():
             raise RuntimeError("An exception occurred in the parameter thread.")
-        logger.info("[trainer] end")
 
 
 # ----------------------------
@@ -655,6 +667,7 @@ def train(mp_cfg: MpConfig, parameter_dat: Optional[Any] = None, memory_dat: Opt
             queue_train_to_mem = mp.Queue()  # put->mp->get
             remote_board: Any = manager.Value(ctypes.c_char_p, None)  # set->th->mp->get
             last_mem_queue = mp.Queue()
+            last_worker_param_queue = mp.Queue()
 
             # params
             remote_board.set(pickle.dumps((0, parameter_dat)))
@@ -669,6 +682,7 @@ def train(mp_cfg: MpConfig, parameter_dat: Optional[Any] = None, memory_dat: Opt
                     remote_board,
                     actor_id,
                     end_signal,
+                    last_worker_param_queue,
                 )
                 ps = mp.Process(target=_run_actor, args=params)
                 actors_ps_list.append(ps)
@@ -731,15 +745,32 @@ def train(mp_cfg: MpConfig, parameter_dat: Optional[Any] = None, memory_dat: Opt
                 if end_signal.value:
                     break
 
-            # params
+            # --- parameter
             dat = remote_board.value
             if dat is not None:
                 _, parameter_dat = pickle.loads(dat)
-
-            # memory
             try:
-                if not last_mem_queue.empty():
-                    memory_dat = last_mem_queue.get(timeout=60 * 10)
+                if context.rl_config.use_update_parameter_from_worker():
+                    work_params_dat = last_worker_param_queue.get(timeout=60 * 30)
+                    logger.info("actor0 parameter recved.")
+                    # tf/torchがimportされる可能性あり
+                    trainer_parameter = context.rl_config.make_parameter()
+                    trainer_parameter.restore(parameter_dat)
+                    worker_parameter = context.rl_config.make_parameter()
+                    worker_parameter.restore(work_params_dat)
+                    trainer_parameter.update_from_worker_parameter(worker_parameter)
+                    parameter_dat = trainer_parameter.backup()
+            except Exception:
+                logger.info(traceback.format_exc())
+                logger.error("Failed to receive parameter data.")
+
+            # --- memory
+            try:
+                if mp_cfg.return_memory_data:
+                    logger.info("memory data wait...")
+                    t0 = time.time()
+                    memory_dat = last_mem_queue.get(timeout=mp_cfg.return_memory_timeout)
+                    logger.info(f"memory data recived. {time.time() - t0:.3f}s")
             except Exception:
                 logger.info(traceback.format_exc())
                 logger.error("Failed to receive memory data.")