+++ tests/quick/base/rl/test_worker_run_raw.py
@@ -45,7 +45,7 @@ class WorkerRunStubEpisode(RLWorker):
         assert False
 
     def on_step(self, worker):
-        if self.step_in_episode == 0:
+        if self.step_in_episode == 1:
             assert worker.state == [1]
             assert worker.next_state == [2]
             if self.config.use_render_image_state():
@@ -57,7 +57,7 @@ class WorkerRunStubEpisode(RLWorker):
             assert worker.terminated == False  # noqa: E712
             assert worker.invalid_actions == [1]
             assert worker.next_invalid_actions == [2]
-        elif self.step_in_episode == 1:
+        elif self.step_in_episode == 2:
             assert worker.state == [2]
             assert worker.next_state == [3]
             if self.config.use_render_image_state():
@@ -205,7 +205,7 @@ class WorkerRunStubEpisodeStacked(RLWorker):
         assert False
 
     def on_step(self, worker):
-        if self.step_in_episode == 0:
+        if self.step_in_episode == 1:
             assert worker.state == [0, 1]
             assert worker.next_state == [1, 2]
             if self.config.use_render_image_state():
@@ -235,7 +235,7 @@ class WorkerRunStubEpisodeStacked(RLWorker):
             assert worker.terminated == False  # noqa: E712
             assert worker.invalid_actions == [1]
             assert worker.next_invalid_actions == [2]
-        elif self.step_in_episode == 1:
+        elif self.step_in_episode == 2:
             assert worker.state == [1, 2]
             assert worker.next_state == [2, 3]
             if self.config.use_render_image_state():