+++ tests/quick/base/run/test_play_mp.py
@@ -30,6 +30,7 @@ def test_actor(mocker: pytest_mock.MockerFixture, interrupt_stop: bool):
     remote_qsize = cast(sharedctypes.Synchronized, mp.Value(ctypes.c_int, 0))
     remote_board = _DummyValue(None)
     end_signal = _DummyValue(False)
+    last_worker_param_queue = mp.Queue()
 
     # --- create task
     c = mocker.Mock(spec=RunCallback)
@@ -64,6 +65,7 @@ def test_actor(mocker: pytest_mock.MockerFixture, interrupt_stop: bool):
         remote_board,
         0,
         end_signal,
+        last_worker_param_queue,
     )
 
     assert end_signal.value